Refactor: Main 페이지 firebase config 호출 제거 및 리펙토링
